back/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
pydantic
python-dotenv
torch
transformers
pillow
